hypercode-core/pyproject.toml
[tool.poetry]
name = "hypercode"
version = "0.1.0"
description = "Programming language for neurodivergent brains. Multi-paradigm with textual + visual syntax."
authors = ["Lyndz Williams <welshdog@example.com>"]
license = "CC-BY 4.0"
readme = "README.md"
repository = "https://github.com/welshDog/hypercode"
keywords = ["programming", "language", "neurodivergent", "quantum", "molecular", "accessibility"]

[tool.poetry.dependencies]
python = ">=3.10"
antlr4-python3-runtime = ">=4.13"
pydantic = ">=2.0"
typing-extensions = ">=4.0"

[tool.poetry.group.quantum.dependencies]
qiskit = { version = ">=0.43", optional = true }
qiskit-aer = { version = ">=0.13", optional = true }

[tool.poetry.group.molecular.dependencies]
# Add molecular computation dependencies here, marked as optional
# For example:
# pennylane = { version = "^0.35.0", optional = true }

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"
pytest-mock = "^3.10"
black = "^23.0"
isort = "^5.12.0"
flake8 = "^6.0"
flake8-bugbear = "^23.0"
flake8-docstrings = "^1.7.0"
mypy = "^1.0"
ipython = "^8.0"
doc8 = "^1.0"
Sphinx = "^5.0.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-autodoc-typehints = "^1.19.0"
myst-parser = "^0.18.0"
sphinxcontrib-mermaid = "^0.8.0"
pre-commit = "^3.0"

[tool.poetry.scripts]
hypercode = "hypercode.cli:main"
hypercode-docs = "scripts.build_docs:main"

[tool.sphinx-docs]
source-dir = "docs/source"
build-dir = "docs/build"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=hypercode --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
